+++ Backend/Agronexis.Model/RequestModel/PriceRequestModel.cs
@@ -15,8 +15,8 @@ namespace Agronexis.Model.RequestModel
         public Guid CurrencyId { get; set; }
         public decimal Price { get; set; }
         public bool IsDiscounted { get; set; }
-        public int DiscountPercentage { get; set; }
-        public int DiscountedAmount { get; set; }
+        public decimal DiscountPercentage { get; set; }
+        public decimal DiscountedAmount { get; set; }
         public string? SkuNumber { get; set; }
         public Guid WeightId { get; set; }
         public string? Barcode { get; set; }
